requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
ccxt==4.4.99
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.0
cryptography==45.0.6
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
exceptiongroup==1.3.0
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.9.4
mt5==0.0.3.1
multidict==6.6.3
multitasking==0.0.12
numpy==2.0.2
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
polars==1.32.2
propcache==0.3.2
protobuf==6.31.1
pycares==4.10.0
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
regex==2025.7.31
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
